removing db testing for now
THe problem is that replication slots are a postmaster setting and
therefore have to be set at database start time.  The best way to
do this would be with a helper script which goes, runs
ALTER SYSTEM and then restarts PostgreSQL.  Until this is done
however we get false failures.  So emulating the cpantesters
environment until then